pyproject.toml
[tool.poetry]
name = "sagemcom_api"
version = "1.0.0"
description = "Python client to interact with SagemCom F@st routers via internal API's."
authors = ["Mick Vleeshouwer <mick@imick.nl>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/iMicknl/python-sagemcom-api"
repository = "https://github.com/iMicknl/python-sagemcom-api"
keywords = ["sagemcom", "f@st"]

packages = [
    { include = "sagemcom_api" }
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/iMicknl/python-sagemcom-api/issues"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
aiohttp = "^3.7.3"
pyhumps = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pre-commit = "^2.9.3"
black = "^20.8b1"
pylint = "^2.6.0"
isort = "^5.7.0"
mypy = "^0.790"
flake8 = "^3.8.4"
pyupgrade = "^2.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# https://github.com/PyCQA/isort/wiki/isort-Settings
profile = "black"
force_sort_within_sections = true
combine_as_imports = true
